Use default_storage instead of io.open
When deploying using a build server, we want the manifest
file to be universally available. To do so, we can use django's
file storage system. The thing is built, the manifest sent
to s3, staticfiles are collected, also sent to s3 and when things
need to be read, they are retrieved through the manifest in s3.